Removed global packets-per-tick rate limit This limit allows attackers to deny service to the server by swamping it out with packets. Packets from blocked addresses still contributed to the packet limit, which effectively makes this useless as an anti-attack measure - instead it just makes the server handicap itself when under attack.
Instead, trust the IP banning feature to prevent attackers overloading the server with traffic. We have to read and discard the packets sooner or later - better sooner so that the server can resume normal operations as soon as possible.

As of a5f22f8c8331719b3fa897824f5ff38c763e4e40 IP blocks are now done in real time, so there is no worry of an address managing to make a server tick take forever with too many packets.