A fix for issues, caused by a wrong MTU size.
I don't how does this work, but when I've changed my router, I've noticed an issue with infinite "Building terrain..." caused by the full packet queue bigger than real client's MTU.
I've tested this a lot, and this "hack" helped. The client won't get a connection reply if it sent a wrong MTU, so it resends a request with a lower MTU size.