First look at per-session inter-thread channels this is a working implementation of #52.
I have my doubts that this gigantic refactor was worth the effort, because there's still lock contention on Snooze to consider; however, this does substantially reduce the problem described in pmmp/pthreads#42, as well as reducing the overhead of inter-thread communication by removing the need to transmit session IDs.

This currently relies on each end of the IPC channels to roundtrip session open/close notifications to setup/cleanup stuff; I'll try to improve this before landing this in main.