Session: Reintroduce delayed sending for disconnect notification
this is directed at MCPE to work around the infamous race condition bug
that causes disconnects and transfers to be unreliable.
It reintroduces the old mechanism for server-initiated disconnects that
was used in 0.12:
- First, flush all packets out to clients
- Send disconnect notification
- Wait for ack
- Discard session

On 0.13, it sent the disconnect notification immediately when disconnect
was initiated, which caused the client to sometimes not handle some
packets.