SendReliabilityLayer: do not send reliable packets outside the client's reliable window
The client implements a window of 512 packets, outside of which packets will be dropped. Eventually, the server will resend them, but after a long delay.
To avoid this, we buffer packets whose reliable message indexes are too large for the current reliable window.
This significantly improves performance when transmitting large amounts of data.