Port code to use ext-encoding
this is a test run to see what it's like to work with the new API.

So far, it seems to be OK, although I keep forgetting the parameters for the static methods, which is annoying.
Instinctively I'm reaching for instance methods, and I have to keep reminding myself why I chose to implement
the new API this way instead of using instance methods (it's because this design is more resistant to mistakes).