Add support for anti spoofing cookies
this is a bit nasty because RakNet's protocol is poorly designed, but it does the job.
A CRC32 is computed from the remote IP, port, and a cryptographically secure random salt
which is known only to the server and is rotated every 5 seconds.
This is similar to the approach taken by gophertunnel.

This approach is preferred over completely random cookies because we
don't have to keep a map of client address -> cookie for verification
this way, which means that spoofed IPs can't flood the server memory
with useless cookies.

This feature may be disabled by setting the rotation interval to 0 in the Server constructor.
Notably, OVH has their own anti-spoofing measure that uses a fake random MTU size, and their
defences are known to not work with these cookie checks as seen with
other projects.